requirements.txt
# Copyright (c) MONAI Consortium
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#     http://www.apache.org/licenses/LICENSE-2.0
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

torch>=1.7
tensorflow==2.8.0
itk>=5.3rc4.post3; python_version == '3.10'
monai[nibabel, skimage, pillow, tensorboard, gdown, ignite, torchvision, itk, tqdm, lmdb, psutil, openslide, fire, mlflow]>=1.1.0
uvicorn==0.17.6
pydantic>=1.8.2
python-dotenv==0.20.0
fastapi==0.78.0
python-multipart==0.0.5
requests-toolbelt==0.9.1
httpx==0.23.0
schedule==1.1.0
dicomweb-client==0.56.2
timeloop==1.0.2
expiringdict==1.2.1
expiring_dict==1.1.0
cachetools==5.2.0
watchdog==2.1.8
pydicom==2.3.0
pydicom-seg==0.4.0
pynetdicom==2.0.2
pynrrd==0.4.3
opencv-python-headless==4.5.5.64
numpymaxflow==0.0.5
girder-client==3.1.14
ninja==1.10.2.3
einops>=0.3.2
PyYAML==6.0
filelock==3.7.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
shapely==1.8.2
requests==2.28.1
scikit-learn
scipy
albumentations
geojson
geopandas
numba

#sudo apt-get install openslide-tools -y
